# Conflicts:
#	server/routes/etl.js

+++ server/routes/etl.js
@@ -81,7 +81,7 @@ wss.register("originalStats", async (message, socket, progress) => {
     await createOrUpdate(projectId, userId, result)
     return { status: 200, message: 'ok', result }
   } catch (e) {
-    console.log({ ...e })
+    console.log({...e})
     let error = e
     if (e.response && e.response.data) error = e.response.data
     return { status: 500, message: 'get index stats failed', error }
@@ -100,29 +100,31 @@ wss.register('newEtl', async (message, socket, process) => {
   if (result.userId !== userId) return { status: 420, message: 'error' }
   const project = result
   const stats = project.stats
-  stats[project.target].isTarget = true
 
-  let deletedValues = []
-  if (project.targetArray && project.targetArray.length > 1) {
-    deletedValues = Object.keys(project.colValueCounts[project.target]).filter(k => !project.targetArray.includes(k))
-  } else {
-    deletedValues = Object.entries(project.colValueCounts[project.target]).sort((a, b) => b[1] - a[1]).map(([k]) => k)
-  }
+  if(project.problemType && project.problemType !=='Clustering' && project.problemType !== 'Outlier' ){
+    stats[project.target].isTarget = true
+    let deletedValues = []
+    if (project.targetArray && project.targetArray.length > 1) {
+      deletedValues = Object.keys(project.colValueCounts).filter(k => !project.targetArray.includes(k))
+    } else {
+      deletedValues = Object.entries(project.colValueCounts).sort((a, b) => b[1] - a[1]).map(([k]) => k)
+    }
 
-  stats[project.target].mapFillMethod = {
-    ...Object.entries(project.renameVariable).reduce((prev, [key, value]) => {
-      prev[key] = {
-        type: 'replace',
-        value
-      }
-      return prev
-    }, {}),
-    ...deletedValues.reduce((prev, key) => {
-      prev[key] = {
-        type: 'delete'
-      }
-      return prev
-    }, {})
+    stats[project.target].mapFillMethod = {
+      ...Object.entries(project.renameVariable).reduce((prev, [key, value]) => {
+        prev[key] = {
+          type: 'replace',
+          value
+        }
+        return prev
+      }, {}),
+      ...deletedValues.reduce((prev, key) => {
+        prev[key] = {
+          type: 'delete'
+        }
+        return prev
+      }, {})
+    }
   }
 
   for (let key in stats) {
@@ -143,17 +145,17 @@ wss.register('newEtl', async (message, socket, process) => {
   return new Promise((resolve, reject) => {
     const interval = setInterval(async () => {
       const { data } = await axios.get(`${esServicePath}/etls/getTaskByOpaqueId/${opaqueId}`)
+      console.log('interval', data.task);
       if (data.task) {
-        if (data.task.status) {
-          const status = data.task.status
-          const progress = 95 * (status.created + status.deleted) / status.total || 0
-          process({ progress, status: 1 })
-        }
-      } else {
+        const status = data.task.status
+        const progress = 95 * (status.created + status.deleted) / status.total || 0
+        process({ progress, status: 1 })
+      }
+      else {
         clearInterval(interval)
         process({ progress: 95, status: 1 })
         const { data: { totalFixedCount, deletedCount } } = await axios.post(`${esServicePath}/etls/${project.originalIndex}/fixedLines`, stats)
-        createOrUpdate(projectId, userId, { etlIndex, opaqueId, totalFixedLines: totalFixedCount, deletedCount, etlProgress: 0 })
+        createOrUpdate(projectId, userId, { etlIndex, opaqueId, totalFixedLines: totalFixedCount, deletedCount })
         resolve({
           status: 200,
           message: 'ok',
